datatable refactored
