+++ share/dashboard_react/src/components/DataTable/styles.module.scss
@@ -1,8 +1,8 @@
 .table {
   overflow-x: auto;
   width: 100%;
-  padding-left: 8px;
-  font-size: 15px;
+  padding-left: rem(8);
+  font-size: rem(15);
   .headerRow {
     background-color: var(--tertiary-color);
 
@@ -13,30 +13,27 @@
       color: var(--text-color) !important;
       border-color: var(--quaternary-color) !important;
     }
-
-    .tableColumn {
-      padding: 2px 4px;
-      border-right: 1px solid;
-      border-color: var(--tertiary-color);
-      border-bottom: none;
-      height: 36px;
-    }
-    .tableColumnEven {
-      background-color: var(--secondary-gray-color);
-    }
-
     .fixedColumn {
       position: sticky;
       left: -16px;
       z-index: 2;
       background-color: var(--primary-color);
     }
-
-    .redBlinking {
-      background-color: var(--danger-primary-color);
-    }
-    .orangeBlinking {
-      background-color: var(--warning-primary-color);
-    }
+  }
+  .tableColumn {
+    padding: rem(2) rem(4);
+    border-right: 1px solid;
+    border-color: var(--tertiary-color);
+    border-bottom: none;
+    height: 36px;
+  }
+  .tableColumnEven {
+    background-color: var(--secondary-gray-color);
+  }
+  .redBlinking {
+    background-color: var(--danger-tertiary-color);
+  }
+  .orangeBlinking {
+    background-color: var(--warning-tertiary-color);
   }
 }
