+++ share/dashboard_react/src/components/Gauge/index.jsx
