+++ share/dashboard_react/src/components/Gauge/styles.module.scss
@@ -18,6 +18,7 @@
     .labelText {
       display: block;
       font-size: rem(12);
+      font-weight: 500;
     }
   }
 }
